requirements.txt
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
html5lib==1.1
firebase-admin==6.6.0
python-telegram-bot==21.6
APScheduler==3.10.4
user-agent==0.1.10
cloudscraper==1.2.71
httpx==0.27.2
